+++ src/types/io.ts
@@ -232,7 +232,6 @@ export type AoGatewayWithAddress = AoGateway & {
 export type AoGatewayDelegate = {
   delegatedStake: number;
   startTimestamp: Timestamp;
-  vaults: Record<WalletAddress, AoVaultData>;
 };
 
 export type AoGatewayDelegateWithAddress = AoGatewayDelegate & {
@@ -401,6 +400,26 @@ export type AoSubmitAuctionBidParams = AoArNSNameParams & {
   quantity?: number;
   type?: 'lease' | 'permabuy';
   years?: number;
+};
+
+export type AoGatewayRegistrySettings = {
+  delegates: {
+    minStake: number;
+    withdrawLengthMs: number;
+  };
+  observers: {
+    tenureWeightDays: number;
+    tenureWeightPeriod: number;
+    maxTenureWeight: number;
+    maxPerEpoch: number;
+  };
+  operators: {
+    minStake: number;
+    withdrawLengthMs: number;
+    leaveLengthMs: number;
+    failedEpochCountMax: number;
+    failedEpochSlashRate: number;
+  };
 };
 
 export type AoGatewayRegistrySettings = {